refactor database

+++ src/main/java/com/fredettly/demo/controller/QuestionController.java
@@ -21,7 +21,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id") Integer id,
+    public String question(@PathVariable(name = "id") Long id,
                            Model model) {
         QuestionDTO questionDTO = questionService.getById(id);
         questionService.incView(id);
@@ -30,7 +30,7 @@ public class QuestionController {
     }
 
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable(name = "id") Integer id
+    public String delete(@PathVariable(name = "id") Long id
     ) {
         questionMapper.deleteByPrimaryKey(id);
         return "redirect:/";
